Make external process task failure more robust.
This ocmmit is intended to make the ExternalProcess tasks more robust with respect to handling failures. As these tasks
represent an external process the only failure cases that really matter are those of the external process. Internal
Python errors that occur during the execution of the external process probably don't reflect the actual state of the
task/item so they should be ignored while the process continues to run.